worker/requirements.txt
celery[redis]>=5.4.0
redis>=5.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pillow>=12.1.0
boto3>=1.35.0
realesrgan>=0.3.0
basicsr>=1.4.2
# Installed by Dockerfile from PyTorch CPU index; keep range open so pre-install is satisfied
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
opencv-python-headless>=4.8.0
pydantic-settings>=2.6.0
rembg[cpu]>=2.0.0
gfpgan>=1.3.8
img2pdf>=0.5.0
pypdf>=4.0.0
pillow-heif>=0.21.0
cairosvg>=2.7.0
pytesseract>=0.3.10
pymupdf>=1.24.0
neuro-deblur>=0.1.1